refactor(headless-styles): move Position type up to shared level
