+++ packages/headless-styles/src/components/types.ts
@@ -27,3 +27,16 @@ export interface InputFieldOptions extends FieldOptions {
 export type CheckboxDirection = 'row' | 'col'
 export type IconSize = 's' | 'm' | 'l'
 export type Tech = 'svelte' | ''
+export type Position =
+  | 'topStart'
+  | 'top'
+  | 'topEnd'
+  | 'rightStart'
+  | 'right'
+  | 'rightEnd'
+  | 'bottomStart'
+  | 'bottom'
+  | 'bottomEnd'
+  | 'leftStart'
+  | 'left'
+  | 'leftEnd'
